+++ tests/run-tests.lisp
@@ -69,6 +69,9 @@
     (run-suite cl-asm/test.m68k-parser)
     (run-suite cl-asm/test.m68k)
 
+    ;; --- Outils ---
+    (run-suite cl-asm/test.acme2clasm)
+
     ;; --- Bilan global ---
     (format t "~%")
     (format t "--- ~A ---~%" (%msg "Core" "Core"))
@@ -121,6 +124,10 @@
     (format t "=== m68k         : ~3D OK, ~D KO~%"
             cl-asm/test.m68k:*pass*
             cl-asm/test.m68k:*fail*)
+    (format t "--- ~A ---~%" (%msg "Outils" "Tools"))
+    (format t "=== acme2clasm   : ~3D OK, ~D KO~%"
+            cl-asm/test.acme2clasm:*pass*
+            cl-asm/test.acme2clasm:*fail*)
     (format t "-------------------------------~%")
     (format t (%msg "=== TOTAL        : ~3D OK, ~D KO sur ~D tests~%"
                     "=== TOTAL        : ~3D OK, ~D KO out of ~D tests~%")